Unify the scalar and interval API CRUD/PATCH harnesses behind one base
The create/get/update/PATCH surface (with the null/value/clear-nullable
semantics) was duplicated across two parallel bases — EntityTests (scalar
wire) and IntervalEntityTests (object wire) — kept in parity only by
convention, which had already drifted once before #77.

Factor the shared scenarios into a single abstract EntityCrudTestsBase,
with the wire-shape differences as abstract members: the value's request
body and its strong-typed form, and the GET read assertion. The two bases
become thin adapters that add only their own invalid-payload cases (a
malformed scalar vs. Start > End / a missing required endpoint). The 27
scalar and 4 interval leaf classes are unchanged.

Verified against both providers via Testcontainers: 1281 passed, 0 failed.
